ShotoverManager -> ShotoverProcess in cassandra_int_tests::cassandra_redis_cache
