+++ test-helpers/Cargo.toml
@@ -31,3 +31,5 @@ redis = { version = "0.22.0", features = ["tokio-comp", "cluster"] }
 tokio-io-timeout = "1.1.1"
 tokio-openssl = "0.6.2"
 inferno = "0.11.13"
+itertools = "0.10.1"
+reqwest = "0.11.6"
